requirements.txt
CherryPy>=10.2.1
jaraco.postgres>=1.6
psycopg2>=2.7.1
SQLAlchemy>=1.1.6
